Palindromes
A palindromic number. Start with any number, reverse its digits, and then add the two numbers. Keep repeating the process and eventually a palindromic number should emerge. We can try it with 87, for example: 87 + 78 = 165; 165 + 561 = 726; 726 + 627 = 1353; 1353 + 3531 = 4884. We found a palindrome in 4 steps. Or again with, say, 159, which only takes two steps: 159 + 951 = 1110; 1110 + 0111 = 1221. The number of steps is unpredictable: try it with 89 and you'll need 24 – to get to 8,813,200,023,188!